+++ halo2_proofs/src/plonk/keygen.rs
@@ -287,26 +287,22 @@ where
             .map(|poly| vk.domain.lagrange_from_vec(poly)),
     );
 
-    let fixed_polys: Vec<_> = fixed
-        .iter()
-        .map(|poly| vk.domain.lagrange_to_coeff(poly.clone()))
-        .collect();
+    let fft_twiddles = vk.domain.proving_key_twiddles();
+    let (fixed_polys, fixed_cosets) = vk
+        .domain
+        .batch_lagrange_to_coeff_and_extended(&fixed, &fft_twiddles);
 
-    let fixed_cosets = fixed_polys
-        .iter()
-        .map(|poly| vk.domain.coeff_to_extended(poly.clone()))
-        .collect();
-
-    let permutation_pk = assembly
-        .permutation
-        .build_pk(params, &vk.domain, &cs.permutation);
+    let permutation_pk =
+        assembly
+            .permutation
+            .build_pk(params, &vk.domain, &cs.permutation, &fft_twiddles);
 
     // Compute l_0(X)
     // TODO: this can be done more efficiently
     let mut l0 = vk.domain.empty_lagrange();
     l0[0] = C::Scalar::ONE;
-    let l0 = vk.domain.lagrange_to_coeff(l0);
-    let l0 = vk.domain.coeff_to_extended(l0);
+    let l0 = vk.domain.lagrange_to_coeff_with_twiddles(l0, &fft_twiddles);
+    let l0 = vk.domain.coeff_to_extended_with_twiddles(l0, &fft_twiddles);
 
     // Compute l_blind(X) which evaluates to 1 for each blinding factor row
     // and 0 otherwise over the domain.
@@ -314,15 +310,23 @@ where
     for evaluation in l_blind[..].iter_mut().rev().take(cs.blinding_factors()) {
         *evaluation = C::Scalar::ONE;
     }
-    let l_blind = vk.domain.lagrange_to_coeff(l_blind);
-    let l_blind = vk.domain.coeff_to_extended(l_blind);
+    let l_blind = vk
+        .domain
+        .lagrange_to_coeff_with_twiddles(l_blind, &fft_twiddles);
+    let l_blind = vk
+        .domain
+        .coeff_to_extended_with_twiddles(l_blind, &fft_twiddles);
 
     // Compute l_last(X) which evaluates to 1 on the first inactive row (just
     // before the blinding factors) and 0 otherwise over the domain
     let mut l_last = vk.domain.empty_lagrange();
     l_last[params.n as usize - cs.blinding_factors() - 1] = C::Scalar::ONE;
-    let l_last = vk.domain.lagrange_to_coeff(l_last);
-    let l_last = vk.domain.coeff_to_extended(l_last);
+    let l_last = vk
+        .domain
+        .lagrange_to_coeff_with_twiddles(l_last, &fft_twiddles);
+    let l_last = vk
+        .domain
+        .coeff_to_extended_with_twiddles(l_last, &fft_twiddles);
 
     Ok(ProvingKey {
         vk,
@@ -333,5 +337,6 @@ where
         fixed_polys,
         fixed_cosets,
         permutation: permutation_pk,
+        fft_twiddles,
     })
 }