+++ halo2_proofs/src/plonk/permutation/prover.rs
@@ -171,11 +171,11 @@ impl Argument {
 
             let permutation_product_commitment_projective = params.commit_lagrange(&z, blind);
             let permutation_product_blind = blind;
-            let z = domain.lagrange_to_coeff(z);
+            let z = domain.lagrange_to_coeff_with_twiddles(z, &pk.fft_twiddles);
             let permutation_product_poly = z.clone();
 
-            let permutation_product_coset =
-                evaluator.register_poly(domain.coeff_to_extended(z.clone()));
+            let permutation_product_coset = evaluator
+                .register_poly(domain.coeff_to_extended_with_twiddles(z.clone(), &pk.fft_twiddles));
 
             let permutation_product_commitment =
                 permutation_product_commitment_projective.to_affine();